+++ src/main/java/org/folio/search/controller/FolioTenantController.java
@@ -1,14 +1,8 @@
 package org.folio.search.controller;
 
-import static org.folio.search.utils.SearchUtils.INSTANCE_RESOURCE;
-
-import java.util.Collection;
-import java.util.stream.Stream;
 import lombok.extern.log4j.Log4j2;
-import org.folio.search.service.IndexService;
 import org.folio.search.service.KafkaAdminService;
 import org.folio.search.service.SearchTenantService;
-import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.controller.TenantController;
 import org.folio.spring.service.TenantService;
 import org.folio.tenant.domain.dto.TenantAttributes;
@@ -22,23 +16,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController("folioTenantController")
 public class FolioTenantController extends TenantController {
 
-  private static final String REINDEX_PARAM_NAME = "runReindex";
-
   private final KafkaAdminService kafkaAdminService;
   private final SearchTenantService tenantService;
-  private final IndexService indexService;
-  private final FolioExecutionContext context;
 
-
-  public FolioTenantController(TenantService baseTenantService,
-                               KafkaAdminService kafkaAdminService, SearchTenantService tenantService,
-                               IndexService indexService, FolioExecutionContext context) {
-
+  public FolioTenantController(TenantService baseTenantService, KafkaAdminService kafkaAdminService,
+                               SearchTenantService tenantService) {
     super(baseTenantService);
     this.kafkaAdminService = kafkaAdminService;
     this.tenantService = tenantService;
-    this.indexService = indexService;
-    this.context = context;
   }
 
   @Override
@@ -50,14 +35,7 @@ public class FolioTenantController extends TenantController {
 
     if (tenantInit.getStatusCode() == HttpStatus.OK) {
       tenantService.initializeTenant();
-
-      Stream.ofNullable(tenantAttributes.getParameters()).flatMap(Collection::stream)
-        .filter(parameter -> parameter.getKey().equals(REINDEX_PARAM_NAME)
-          && Boolean.parseBoolean(parameter.getValue()))
-        .findFirst().ifPresent(parameter -> {
-          indexService.dropIndex(INSTANCE_RESOURCE, context.getTenantId());
-          indexService.createIndex(INSTANCE_RESOURCE, context.getTenantId());
-        });
+      tenantService.reIndexInstances(tenantAttributes);
     }
 
     log.info("Tenant init has been completed [response={}]", tenantInit);
