+++ src/main/java/org/folio/search/service/SearchTenantService.java
@@ -1,5 +1,7 @@
 package org.folio.search.service;
 
+import java.util.Collection;
+import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.search.configuration.properties.SearchConfigurationProperties;
@@ -7,12 +9,14 @@ import org.folio.search.domain.dto.LanguageConfig;
 import org.folio.search.model.SearchResource;
 import org.folio.search.service.systemuser.SystemUserService;
 import org.folio.spring.FolioExecutionContext;
+import org.folio.tenant.domain.dto.TenantAttributes;
 import org.springframework.stereotype.Service;
 
 @Log4j2
 @Service
 @RequiredArgsConstructor
 public class SearchTenantService {
+  private static final String REINDEX_PARAM_NAME = "runReindex";
 
   private final IndexService indexService;
   private final FolioExecutionContext context;
@@ -45,5 +49,12 @@ public class SearchTenantService {
 
       indexService.dropIndex(resource.getName(), context.getTenantId());
     }
+  }
+
+  public void reIndexInstances(TenantAttributes tenantAttributes) {
+    Stream.ofNullable(tenantAttributes.getParameters()).flatMap(Collection::stream)
+      .filter(parameter -> parameter.getKey().equals(REINDEX_PARAM_NAME)
+        && Boolean.parseBoolean(parameter.getValue()))
+      .findFirst().ifPresent(parameter -> indexService.reindexInventory(context.getTenantId(), null));
   }
 }
